Fix command list search bug, simplify mentions
There was an issue that e.g. a command !foobar would be treated as
matching !foo when searching the commands list to check for clashes when
adding a command, or when removing a command. This has now been corrected.

Additionally, mentions were being added in an overly roundabout way using
a guildMember object isntead of a User object. This has been corrected.