Fix command list search bug, use ES6 functions
There was an issue that e.g. a command !foobar would be treated as
matching !foo when searching the commands list to check for clashes when
adding a command, or when removing a command. This has now been corrected.

Additionally, use of indexOf has been replaced by appropriate ES6
functions.