Fix bug where newcommand makes dodgy commands
Due to a very silly oversight, the newcommand command was adding Command
objects whose command property was an array containing a string, rather
than a plain string. This worked in some contexts where it was converted
to a string, but failed badly when attempting to call string methods on
it.