Fix commands running own substrings
With the original function to check whether a message is a command,
a match would be detected not just on a true command, but any other
command whose call is a substring starting at the beginning of the
one actually being called would also be called. This adds a simple
workaround so only the intended command is actually called (as long as
commands do not have spaces in them).