Fix bugs in Menus module
The menus module now works correctly to allow back-and-forth between
the bot and the user. Other modules may require() the menus module, at
which point they can call Menus.present() to get a Promise for the
answer to some query they want to make of the user from a list of
options.

This commit fixes a few incorrect method calls and typoes that were
causing the Menus module to fail.